Reorganised files
Partials contains parts of pages (nav, footer, cards)
Routes contains the pages to be rendered at different routes (home, movies, generator...)
