+++ src/routes/Generator.js
@@ -1,0 +1,7 @@
+export default function Generator() {
+	return (
+		<div id="generator">
+			<h1>This is a generator</h1>
+		</div>
+	);
+}
